refactor: pages 레이어 ui/index.ts 파일 제거 및 페이지 라우터 import 경로 수정
