+++ app/note/edit/[noteId]/page.tsx
@@ -1,4 +1,4 @@
-import { NoteEditPage } from '@/(pages)/note/ui';
+import { NoteEditPage } from '@/(pages)/note/ui/NoteEditPage';
 
 export default function Page() {
   return <NoteEditPage />;
