+++ app/note/page.tsx
@@ -1,4 +1,4 @@
-import { NoteListPage } from '@/(pages)/note/ui';
+import { NoteListPage } from '@/(pages)/note/ui/NoteListPage';
 
 export default function Page() {
   return <NoteListPage />;
