--- src/(pages)/note/ui/index.ts
@@ -1,3 +1,0 @@
-export { NoteEditPage } from './NoteEditPage';
-export { NoteListPage } from './NoteListPage';
-
